Adding files to send to database now happens in single transactional action
